+++ tests/test_models.py
@@ -1,32 +1,49 @@
 from django.db import models
+from django.forms import ModelForm
 from django.test import TestCase
 from django.utils.safestring import SafeString
 from django_nh3.models import Nh3Field
 
 
 class Nh3Content(models.Model):
-    """Bleach test model"""
+    """NH3 test model"""
 
-    CHOICES = (("f", "first choice"), ("s", "second choice"))
     content = Nh3Field(
         strip_comments=True,
     )
-    choice = Nh3Field(choices=CHOICES)
     blank_field = Nh3Field(blank=True)
     null_field = Nh3Field(blank=True, null=True)
 
 
-class Nh3NullableContent(models.Model):
-    """Bleach test model"""
+class Nh3ContentModelForm(ModelForm):
+    """NH3 test model form"""
 
+    class Meta:
+        model = Nh3Content
+        fields = ["content"]
+
+
+class Nh3NullableContent(models.Model):
+    """NH3 test model"""
+
+    CHOICES = (("f", "first choice"), ("s", "second choice"))
+    choice = Nh3Field(choices=CHOICES, blank=True)
     content = Nh3Field(blank=True, null=True)
+
+
+class Nh3NullableContentModelForm(ModelForm):
+    """NH3 test model form"""
+
+    class Meta:
+        model = Nh3NullableContent
+        fields = ["choice"]
 
 
 class TestNh3ModelField(TestCase):
     """Test model field"""
 
     def test_cleaning(self):
-        """Test values are bleached"""
+        """Test values are sanitized"""
         test_data = {
             "html_data": "<h1>Heading</h1>",
             "no_html": "Heading",
@@ -68,7 +85,7 @@ class TestNh3NullableModelField(TestCase):
     """Test model field"""
 
     def test_cleaning(self):
-        """Test values are bleached"""
+        """Test values are sanitized"""
         test_data = {
             "none": None,
             "empty": "",
@@ -85,3 +102,45 @@ class TestNh3NullableModelField(TestCase):
         for key, value in test_data.items():
             obj = Nh3NullableContent.objects.create(content=value)
             self.assertEqual(obj.content, expected_values[key])
+
+
+class TestNh3ModelFormField(TestCase):
+    """Test model form field"""
+
+    def test_cleaning(self):
+        """Test values are sanitized"""
+        test_data = {
+            "html_data": "<h1>Heading</h1>",
+            "no_html": "Heading",
+            "spacing": " Heading ",
+        }
+        expected_values = {
+            "html_data": "Heading",
+            "no_html": "Heading",
+            "spacing": "Heading",
+        }
+
+        for key, value in test_data.items():
+            form = Nh3ContentModelForm(data={"content": value})
+            self.assertTrue(form.is_valid())
+            obj = form.save()
+            self.assertEqual(obj.content, expected_values[key])
+
+    def test_stripped_comments(self):
+        """Content field strips comments so ensure they aren't allowed"""
+
+        self.assertFalse(
+            Nh3ContentModelForm(
+                data={"content": "<!-- this is a comment -->"}
+            ).is_valid()
+        )
+
+    def test_field_choices(self):
+        """Content field strips comments so ensure they aren't allowed"""
+        test_data = dict(Nh3NullableContent.CHOICES)
+
+        for key, value in test_data.items():
+            form = Nh3NullableContentModelForm(data={"choice": key})
+            self.assertTrue(form.is_valid())
+            obj = form.save()
+            self.assertEqual(obj.get_choice_display(), value)